Stop using generics for math. it's too much of a pain for android performance

+++ src/test/scala/com/github/gigurra/math/InterpolatorSpec.scala
@@ -14,23 +14,23 @@ class InterpolatorSpec
   "Interpolator" should {
 
     "perp half" in {
-      Interpolator(range = 3.0, perp(_:Double, _:Double, exponent = 0.5, 0.0, 2.0, clamp = false)).interpolate(1.5) shouldBe math.sqrt(2.0)
+      Interpolator(range = 3.0f, perp(_:Float, _:Float, exponent = 0.5f, 0.0f, 2.0f, clamp = false)).interpolate(1.5f) shouldBe math.sqrt(2.0f).toFloat
     }
 
     "be chained" in {
 
-      val i1 = Interpolator(range = 3.0, perp(_:Double, _:Double, exponent = 0.5, 0.0, 2.0, clamp = true))
-      val i2 = Interpolator(range = 3.0, lerp(_:Double, _:Double, 0.0, 2.0, clamp = false))
+      val i1 = Interpolator(range = 3.0f, perp(_:Float, _:Float, exponent = 0.5f, 0.0f, 2.0f, clamp = true))
+      val i2 = Interpolator(range = 3.0f, lerp(_:Float, _:Float, 0.0f, 2.0f, clamp = false))
 
       val ic = i1.andThen(i2).andThen(i1)
 
-      ic.interpolate(-10.0) shouldBe 0.0
-      ic.interpolate(0.0) shouldBe 0.0
-      ic.interpolate(1.5) shouldBe math.sqrt(2.0)
-      ic.interpolate(4.5) shouldBe 1.0
-      ic.interpolate(7.5) shouldBe math.sqrt(2.0)
-      ic.interpolate(9.0) shouldBe 2.0
-      ic.interpolate(10.5) shouldBe 2.0
+      ic.interpolate(-10.0f) shouldBe 0.0f
+      ic.interpolate(0.0f) shouldBe 0.0f
+      ic.interpolate(1.5f) shouldBe math.sqrt(2.0f).toFloat
+      ic.interpolate(4.5f) shouldBe 1.0f
+      ic.interpolate(7.5f) shouldBe math.sqrt(2.0f).toFloat
+      ic.interpolate(9.0f) shouldBe 2.0f
+      ic.interpolate(10.5f) shouldBe 2.0f
     }
 
   }
